Clear the whole structure instead the size of a pointer
The type of "bak" and "pattern" is uint8. The compiler does not have
a way to know how long this array will be. Use data_size from allocation.

Some newer versions may have this fixed.

TODO there will be probably more bugs of this type in the code. Check it.
TODO^2 change the allocation to support zeroing?

Signed-off-by: Petr Cvek <petrcvekcz@gmail.com>